% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{collapseVector}
\alias{collapseVector}
\alias{strcoll}
\title{Coerce a vector to a single character string}
\usage{
collapseVector(x = vector())

strcoll(x = vector())
}
\arguments{
\item{x}{\verb{[NULL | vector()]}

A so-called \emph{generalized} vector: either \code{NULL}, an
\link[base:vector]{atomic vector} or a \link[base:list]{list}.}
}
\value{
A \code{character(1)}.
}
\description{
Create a character string from the values of a vector or a list.
This function is a shortcut to \code{paste0(x, collapse = "")}.
}
\details{
Function \code{\link[=strcoll]{strcoll()}} is an alias to its semantic counterpart
\code{\link[=collapseVector]{collapseVector()}}. It fits package \pkg{base}'s naming conventions
for functions that operate with and on strings.
}
