% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getDotsAsVector}
\alias{.getDotsAsVector}
\title{Destructure parameters passed to ...}
\usage{
.getDotsAsVector(...)
}
\arguments{
\item{...}{\verb{[any]}

Any \R object(s).}
}
\value{
A vector of a type given by the highest one when following \R's
usual \link[base:c]{type hierarchy}. Names are preserved. \code{NULL} is returned
if the resulting vector is empty.
}
\description{
Destructure all values passed to ... and return a (possibly named)
vector containing all values passed directly and/or indirectly to
...
}
\keyword{internal}
