% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simplifyInteger}
\alias{.simplifyInteger}
\title{Simplify an integer vector}
\usage{
.simplifyInteger(int = integer())
}
\arguments{
\item{int}{\verb{[integer()]}

An integer vector.}
}
\value{
An integer vector with a length no greater than the length
of \code{int}.
}
\description{
Remove duplicates from an integer vector and sort the remaining values
in ascending order. This function is optimized for small vectors.
}
\keyword{internal}
