% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introspectors.R
\name{internal-is}
\alias{internal-is}
\alias{.is}
\title{Generic introspector}
\usage{
.is(x, isCb, targetLength = NULL, acceptNA = TRUE)
}
\arguments{
\item{x}{\verb{[any]}
Object to be tested.}

\item{isCb}{\verb{[function]}
Callback function. Example: is.logical(). This is an internal parameter
not visible by users.}

\item{acceptNA}{\verb{[logical(1)]}
Can \code{x} contain NA values?}

\item{targetlength}{\verb{[NULL | integer(1) | double(1)]}
Desired length of \code{x}. No length is enforced if this parameter is NULL.}
}
\value{
A logical(1).
}
\description{
Check that x is of a type given by a callback function. Optionally,
enforce a specific length and accept or reject NA values.
}
\keyword{internal}
