% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocli-package.R
\docType{package}
\name{nanocli-package}
\alias{nanocli}
\alias{nanocli-package}
\title{nanocli: Create command-line interfaces easily}
\description{
Convey messages with style and minimal code. Throw, register, and retrieve errors more easily. Interpolate and format strings with a clean syntax. Avoid dependency hell: this package only has one.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeanmathieupotvin/nanocli}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/nanocli/issues/new}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jm@potvin.xyz} (\href{https://orcid.org/0000-0002-8237-422X}{ORCID})

}
