% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotools-package.R
\docType{package}
\name{nanotools-package}
\alias{nanotools}
\alias{nanotools-package}
\title{nanotools: Simpler Alternatives to R Features}
\description{
A toolbox of simpler alternatives and wrapper functions to \R
features. The intent is threefold:
\enumerate{
\item simplify features whenever possible (keep it simple),
\item minimize reliance on dependencies (keep it small), and
\item provide pure \R alternatives that are as fast and efficient as
possible (keep it fast).
}
}
\section{Context}{
 \R is a very mature language. Its official repository
(\href{https://cran.r-project.org/}{\emph{The Comprehensive R Archive Network}}),
contains more than 16 000 packages. Major communities have emerged over the
years and organized themselves around recognized ecosystems. Each of these
can be considered as meta-languages derived from \R. Some of them greatly
differ from \pkg{base} \R. Others are so complex (dozens of dependencies)
that they have a non-negligible \emph{entry cost}.

While these advanced ecosystems are \emph{powerful}, they can introduce a
non-negligible overhead. It takes time to learn these micro-languages.
It takes time to maintain them in production environments. Sometimes,
there is \emph{no} time. Sometimes, you just need to \emph{get the job done}.
And fast.

Package \pkg{nanotools} seeks to be a lighter alternative that is easy
to learn, use, and manage.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeanmathieupotvin/nanotools}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/nanotools/issues/new}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jm@potvin.xyz} (\href{https://orcid.org/0000-0002-8237-422X}{ORCID})

}
